requirements.txt
requests>=2.31.0
pandas>=2.1.0
sqlite-utils>=3.35
streamlit>=1.29.0
plotly>=5.18.0
folium>=0.15.0
streamlit-folium>=0.15.0
gspread>=5.12.0
google-auth>=2.23.0
openpyxl>=3.1.0
